Update reading from files instead of DB
Annotation data is read from --a=tabSeparatedFile.tsv, with the
strcuture identical to the downloaded one from UCSC browser.
Peaks calling data is read from --in=macs2GeneratedResults.xls.
Output is saved to --out=tabSeparatedFile.tsv with the structure
identical to experiments.uid_islands table.
Parameters left wo changes: --promoter, --log, --upstream.
Note: while testing the following bug was found. If use QSet with
the pointers the order of items is random, so when few or more
isoforms (refseq_id) are found for the same island, txStart and
txEnd may vary from run to run, depending on wich pointer is first
in a QSet.