Fix bug with txStart, txEnd for multiple isoforms
When two or more isoforms assigned for one island,
QSet <annotationPtr> contains random set of ptr,
with the order that depends on memory cell address.
When we set txStart, txEnd we choose the first item
from QSet, so it differs from run to run. To solve
this problem QSet is updated to custom class, that
stores all data in QMultiMap <QString, annotationPtr>
where key is concatenated value of txStart and txEnd.
This allows to maintain a consistent order of items
in a container from run to run.